--refactor: Made a services folder to have all different service functionality at one place.
